+++ src/reference.core.ts
@@ -23,7 +23,7 @@
 import { type Lazy, map } from "@metreeca/core";
 import { equals, immutable } from "@metreeca/core/structures";
 import { isIRI } from "@metreeca/core/resource";
-import { isReference } from "@metreeca/qest";
+import { isReference } from "@metreeca/qest/resource";
 import { all, array, domain, test, type Trace, TraceError, type, values as contains } from "@metreeca/core/trace";
 import { type Scope } from "./index.core.js";
 import type { ReferenceShape } from "./reference.js";