+++ src/string.ts
@@ -128,7 +128,7 @@ import { isRegExp, isString } from "@metreeca/core";
 import { xsd } from "@metreeca/core/datatype";
 import { immutable } from "@metreeca/core/structures";
 import { type Variant } from "@metreeca/core/resource";
-import type { Reference } from "@metreeca/qest";
+import type { Reference } from "@metreeca/qest/resource";
 
 import { TraceError } from "@metreeca/core/trace";
 import { checkString } from "./string.core.js";